+++ agda2rust.cabal
@@ -27,7 +27,11 @@ common warnings
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Agda.Compiler.Rust.Backend Paths_agda2rust
+  exposed-modules:     Agda.Compiler.Rust.Backend
+                       Agda.Compiler.Rust.CommonTypes
+                       Agda.Compiler.Rust.PrettyPrintingUtils
+                       Agda.Compiler.Rust.ToRustCompiler
+                       Paths_agda2rust
   autogen-modules:     Paths_agda2rust
   build-depends:       base >= 4.10 && < 4.20,
                        Agda >= 2.6.4 && < 2.6.5,
