+++ src/Agda/Compiler/Rust/CommonTypes.hs
@@ -1,0 +1,10 @@
+module Agda.Compiler.Rust.CommonTypes (
+  Options(..),
+  CompiledDef,
+  ModuleEnv ) where
+
+data Options = Options { optOutDir :: Maybe FilePath }
+
+type CompiledDef = String
+
+type ModuleEnv = ()
