+++ src/Agda/Compiler/Rust/PrettyPrintingUtils.hs
@@ -1,0 +1,18 @@
+module Agda.Compiler.Rust.PrettyPrintingUtils (
+  bracket,
+  indent,
+  exprSeparator,
+  defsSeparator
+) where
+
+bracket :: String -> String
+bracket str = "{\n" <> str <> "\n}"
+
+indent :: String
+indent = "  "
+
+exprSeparator :: String
+exprSeparator = " "
+
+defsSeparator :: String
+defsSeparator = "\n"
